Add files via upload
What is being committed resolves the issues I created.

The png file is an image of the current database schema. It is subject to change based of what the client wants includes once we receive that information.

The word document lists the fields and has information on each of them include the data type and a description of what it stores.